requirements.txt
torch==1.9.0+cu111
numpy
pillow